fix: verify if is a secundary project before call flows in sectors and queues creation

+++ chats/apps/api/v1/sectors/viewsets.py
@@ -78,38 +78,44 @@ class SectorViewset(viewsets.ModelViewSet):
 
         project = Project.objects.get(uuid=instance.project.uuid)
 
-        is_principal = project.config and project.config.get("its_principal", False)
+        should_use_integration = (
+            project.config
+            and project.config.get("its_principal", False)
+            and instance.secondary_project
+        )
 
-        if is_principal:
-            integrate_use_case = IntegratedTicketers()
-            integrate_use_case.integrate_individual_ticketer(
-                project, instance.secondary_project
-            )
-        elif settings.USE_WENI_FLOWS:
-            content = {
-                "project_uuid": str(instance.project.uuid),
-                "name": instance.name,
-                "config": {
-                    "project_auth": str(instance.external_token.pk),
-                    "sector_uuid": str(instance.uuid),
+        if settings.USE_WENI_FLOWS:
+            if not should_use_integration:
+                content = {
                     "project_uuid": str(instance.project.uuid),
-                    "project_name_origin": instance.name,
-                },
-            }
+                    "name": instance.name,
+                    "config": {
+                        "project_auth": str(instance.external_token.pk),
+                        "sector_uuid": str(instance.uuid),
+                        "project_uuid": str(instance.project.uuid),
+                        "project_name_origin": instance.name,
+                    },
+                }
 
-            flows_client = FlowRESTClient()
-            response = flows_client.create_ticketer(**content)
-            if response.status_code not in [
-                status.HTTP_200_OK,
-                status.HTTP_201_CREATED,
-            ]:
-                instance.delete()
+                flows_client = FlowRESTClient()
+                response = flows_client.create_ticketer(**content)
+                if response.status_code not in [
+                    status.HTTP_200_OK,
+                    status.HTTP_201_CREATED,
+                ]:
+                    instance.delete()
 
-                raise exceptions.APIException(
-                    detail=(
-                        f"[{response.status_code}] Error posting the sector/ticketer on flows. "
-                        f"Exception: {response.content}"
+                    raise exceptions.APIException(
+                        detail=(
+                            f"[{response.status_code}] Error posting the sector/ticketer on flows. "
+                            f"Exception: {response.content}"
+                        )
                     )
+
+            else:
+                integrate_use_case = IntegratedTicketers()
+                integrate_use_case.integrate_individual_ticketer(
+                    project, instance.secondary_project
                 )
 
     def update(self, request, *args, **kwargs):
